+++ App/assets/icon/index.ts
@@ -1,1 +1,5 @@
 export { default as calendar } from './calendar.svg';
+export { default as store } from './store.svg';
+export { default as note } from './note.svg';
+export { default as code } from './code.svg';
+export { default as settings } from './settings.svg';
