+++ App/screens/bottom_tab/TabButton.tsx
@@ -1,0 +1,75 @@
+import { Pressable, StyleSheet, Text, View } from 'react-native';
+import React, { useEffect } from 'react';
+import { BottomTabBarButtonProps } from '@react-navigation/bottom-tabs';
+import SvgIcon from '../../components/SvgIcon';
+import Animated, { useSharedValue } from 'react-native-reanimated';
+import { useAnimatedStyle, withTiming, Easing } from 'react-native-reanimated';
+import haptic from '../../util/haptic';
+import themeChange from '../../util/theme';
+
+const TabButton = (props: any) => {
+  const { item, onPress, accessibilityState } = props;
+  const focused = accessibilityState?.selected;
+  const activeColor = focused ? '#FFF' : '#6B6F78';
+
+  const themeMode = themeChange();
+  const scaleAni = useSharedValue(1);
+  const backgound = useSharedValue(themeMode.secondary);
+
+  useEffect(() => {
+    //테마변경이 될 때마다 기본값 변경해줘야합니다.
+    backgound.value = themeMode.secondary;
+  }, [themeMode]);
+
+  const animatedStyles = useAnimatedStyle(() => {
+    return {
+      backgroundColor: backgound.value,
+      transform: [
+        {
+          scale: withTiming(scaleAni.value, {
+            duration: 200,
+            easing: Easing.bezier(0.25, 0.1, 0.25, 1)
+          })
+        }
+      ]
+    };
+  }, []);
+
+  return (
+    <Pressable
+      style={{
+        alignItems: 'center',
+        justifyContent: 'center',
+        flex: 1
+      }}
+      onPress={onPress}
+      onPressIn={() => {
+        haptic('impactMedium');
+        (scaleAni.value = 0.95), (backgound.value = themeMode.primary);
+      }}
+      onPressOut={() => ((scaleAni.value = 1), (backgound.value = themeMode.secondary))}
+    >
+      <Animated.View
+        style={[
+          {
+            alignItems: 'center',
+            justifyContent: 'center',
+            flex: 1,
+            borderRadius: 10,
+            paddingHorizontal: 15
+          },
+          animatedStyles
+        ]}
+      >
+        <SvgIcon name={item.icon} color={activeColor} />
+        <Text style={{ fontSize: 12, color: activeColor, fontWeight: '700', marginTop: 5 }}>
+          {item.label}
+        </Text>
+      </Animated.View>
+    </Pressable>
+  );
+};
+
+export default TabButton;
+
+const styles = StyleSheet.create({});
